% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{as.partial}
\alias{as.partial}
\title{Convert text or file to a partial}
\usage{
as.partial(input = NULL, text = NULL, knit_meta = list())
}
\arguments{
\item{input}{if you specify a file path here, it will be read in before being passed to knitr (to avoid a working directory mess)}

\item{text}{passed to \code{\link[knitr:knit_child]{knitr::knit_child()}}}

\item{knit_meta}{you can pass a path to figures and other resources here}
}
\description{
This adds the \code{knit_asis} class to a markdown chunk, so that it can be rendered
in the viewer and simply echoed in other knitr chunks. Won't preserve figures
unless the path happens to be the same or you explicitly pass it to the knit_meta argument.
}
