% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlarge_plot.R
\name{enlarge_plot}
\alias{enlarge_plot}
\title{Generate a small plot that will be enlarged in a modal when clicked}
\usage{
enlarge_plot(plot, large_plot = plot, plot_name = NULL,
  width_small = 2, height_small = 2, width_large = 7,
  height_large = 7)
}
\arguments{
\item{plot}{a ggplot2 plot}

\item{large_plot}{a larger version of the same ggplot2 plot. defaults to the first plot if left empty}

\item{plot_name}{optional: specify a meaningful plot name (needs to be unique in the document)}

\item{width_small}{width for the small plot}

\item{height_small}{height for the small plot}

\item{width_large}{width for the large plot}

\item{height_large}{height for the large plot}
}
\description{
Generate a small plot that will be enlarged in a modal when clicked
}
\examples{
# will generate figure in a temporary directory
knit_to_temp_dir()
enlarge_plot(ggplot2::qplot(1:10))
}
