% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_sets.R
\name{join_sets}
\alias{join_sets}
\title{Generate a small plot that will be enlarged in a modal when clicked}
\usage{
join_sets(x, y, by = NULL, ...)
}
\arguments{
\item{x}{a dataset}

\item{y}{another dataset}

\item{by}{variables you plan to merge on}

\item{...}{passed to dplyr join functions}
}
\description{
Generate a small plot that will be enlarged in a modal when clicked
}
\examples{
# will generate figures in a temporary directory
if (requireNamespace("dplyr", quietly = TRUE)) {
   band_members \%>\% join_sets(band_instruments)
}
}
