% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_child_debug.R
\name{knit_child_debug}
\alias{knit_child_debug}
\title{Get some debugging information on various potential problems when making partials}
\usage{
knit_child_debug(...)
}
\arguments{
\item{...}{passed to \code{\link[=partial]{partial()}}}
}
\value{
Returns markdown/HTML text with class "knit_asis"
}
\description{
Get some debugging information on various potential problems when making partials
}
\examples{
if(!requireNamespace("pkgdown", quietly = TRUE) || !pkgdown::in_pkgdown()) {
knit_child_debug()
}
}
