% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_helpers.R
\name{knit_to_temp_dir}
\alias{knit_to_temp_dir}
\title{Knit to temp dir}
\usage{
knit_to_temp_dir()
}
\description{
Use in examples of rmarkdown partials, so they do not clutter the user directory
}
\examples{
knit_to_temp_dir()
enlarge_plot(ggplot2::qplot(1:10))
}
