% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_helpers.R
\name{print.knit_asis}
\alias{print.knit_asis}
\title{Print \code{knit_asis} as rendered HTML in the viewer}
\usage{
\method{print}{knit_asis}(x, ...)
}
\arguments{
\item{x}{the knit_asis object}

\item{...}{ignored}
}
\value{
Invisibly returns its input, either prints its input or sends it to a viewer, if one is defined
}
\description{
Print \code{knit_asis} as rendered HTML in the viewer
}
\examples{
text <- paste(c("### Headline",
"Text"), collapse = "\n")
print(knitr::asis_output(text))
}
