% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diagnostics.R
\name{regression_diagnostics}
\alias{regression_diagnostics}
\title{Show the estimated coefficients in a regression and diagnostics}
\usage{
regression_diagnostics(regression, ...)
}
\arguments{
\item{regression}{an lm object}

\item{...}{passed to \code{\link[=partial]{partial()}}}
}
\description{
Show the estimated coefficients in a regression and diagnostics
}
\examples{
# will generate figures in a temporary directory
data("ChickWeight")
regression <- lm(weight ~ Time, data = ChickWeight)
regression_diagnostics(regression)
}
