% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{as.partial}
\alias{as.partial}
\title{Convert text or file to a partial}
\usage{
as.partial(text = NULL, knit_meta = list())
}
\arguments{
\item{text}{will be returned with the class "knit_asis"}

\item{knit_meta}{you can pass a path to figures and other resources here}
}
\value{
Returns its input as text with class "knit_asis"
}
\description{
This adds the \code{knit_asis} class to a markdown chunk, so that it can be rendered
in the viewer and simply echoed in other knitr chunks. Won't preserve figures
unless the path happens to be the same or you explicitly pass it to the knit_meta argument.
}
\examples{
my_partial <- as.partial("## Headline
Text")
}
