% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_child_debug.R
\name{knit_child_debug}
\alias{knit_child_debug}
\title{Get some debugging information on various potential problems when making partials}
\usage{
knit_child_debug(...)
}
\arguments{
\item{...}{passed to \code{\link[=partial]{partial()}}}
}
\description{
Get some debugging information on various potential problems when making partials
}
\examples{
\dontrun{
knit_child_debug()
}
}
