% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{partial}
\alias{partial}
\title{Knit a child document and output as is (render markup)}
\usage{
partial(input = NULL, ..., text = NULL, output = NULL,
  quiet = TRUE, options = NULL, envir = parent.frame(),
  name = NULL, show_code = FALSE, use_strings = TRUE)
}
\arguments{
\item{input}{if you specify a file path here, it will be read in before being passed to knitr (to avoid a working directory mess)}

\item{...}{ignored, but you can use it to clarify which variables will be used in the rmd partial}

\item{text}{passed to \code{\link[knitr:knit_child]{knitr::knit_child()}}}

\item{output}{if you specify a file path here, where to put the file}

\item{quiet}{passed to \code{\link[knitr:knit_child]{knitr::knit_child()}}}

\item{options}{defaults to NULL.}

\item{envir}{passed to \code{\link[knitr:knit_child]{knitr::knit_child()}}}

\item{name}{a name to use for cacheing and figure paths. Randomly generated if left unspecified.}

\item{show_code}{whether to print the R code for the partial or just the results (sets the chunk option echo = FALSE while the chunk is being rendered)}

\item{use_strings}{whether to read in the child file as a character string (solves working directory problems but harder to debug)}
}
\description{
This slightly modifies the \code{\link[knitr:knit_child]{knitr::knit_child()}} function to have different defaults.
\itemize{
\item the environment defaults to the calling environment.
\item the output receives the class \code{knit_asis}, so that the output will be rendered "as is" by knitr when calling inside a chunk (no need to set \code{results='asis'} as a chunk option).
\item defaults to \code{quiet = TRUE}
}
}
\details{
Why default to the calling environment? Typically this function defaults to the global environment. This makes sense if you want to use knit children in the same context as the rest of the document.
However, you may also want to use knit children inside functions to e.g. summarise a regression using a set of commands (e.g. plot some diagnostic graphs and a summary for a regression nicely formatted).

Some caveats:
\itemize{
\item the function has to return to the top-level. There's no way to \code{\link[=cat]{cat()}} this from loops or an if-condition without without setting \code{results='asis'}. You can however concatenate these objects with \code{\link[=paste.knit_asis]{paste.knit_asis()}}
}
}
\examples{
# super simple partial example
partial(text = "Test")

\dontrun{
# an example of a wrapper function that calls partial with an argument
# ensures distinct paths for cache and figures, so that these calls can be looped in parallel
regression_summary <- function(regression) {
   partial("_regression_summary.Rmd")
}
}
}
