% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_diagnostics.R
\name{regression_diagnostics}
\alias{regression_diagnostics}
\title{Show the estimated coefficients in a regression and diagnostics}
\usage{
regression_diagnostics(regression, ...)
}
\arguments{
\item{regression}{an lm object}

\item{...}{passed to \code{\link[=partial]{partial()}}}
}
\value{
Returns markdown/HTML text with class "knit_asis"
}
\description{
Show the estimated coefficients in a regression and diagnostics
}
\examples{
\dontrun{
# will generate figures in a temporary directory
if(!requireNamespace("pkgdown", quietly = TRUE) || !pkgdown::in_pkgdown()) {
data("ChickWeight")
regression <- lm(weight ~ Time, data = ChickWeight)
regression_diagnostics(regression)
}
}
}
